server/requirements.txt
asgiref==3.10.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
MarkupSafe==3.0.3
pycparser==2.23
PyJWT==2.10.1
python-http-client==3.3.7
sendgrid==6.12.5
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.4
